# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: RedCart Copilot API
  version: 0.2.0
  description: 面向内容电商购物、商家经营与 AI Copilot 的可运行 MVP 接口契约。
servers:
  - url: http://127.0.0.1:18080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
...
    CheckoutItem:
      type: object
      required: [sku_id, quantity]
      properties:
        sku_id:
          type: integer
          format: int64
        quantity:
          type: integer
          minimum: 1
    CheckoutRequest:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CheckoutItem"
        receiver_name:
          type: string
        receiver_phone:
          type: string
        receiver_address:
          type: string
...
    RefundRequest:
      type: object
      properties:
        reason:
          type: string
...
security:
  - bearerAuth: []
paths:
...
  /api/orders:
    get:
      summary: 获取当前消费者订单列表
      responses:
        "200":
          description: 返回订单列表
    post:
      summary: 使用幂等键创建订单
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "201":
          description: 订单创建成功，或命中幂等返回已有订单
        "409":
          description: 业务冲突或库存冲突
  /api/orders/{id}:
    get:
      summary: 获取订单详情
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: 返回订单详情
  /api/orders/{id}/pay:
    post:
      summary: 模拟支付
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: 订单已支付
  /api/orders/{id}/cancel:
    post:
      summary: 取消未支付订单
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: 订单已取消
  /api/orders/{id}/finish:
    post:
      summary: 确认收货并完成订单
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: 订单已完成
  /api/orders/{id}/refund:
    post:
      summary: 发起退款申请
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefundRequest"
      responses:
        "202":
          description: 退款申请已提交
...
  /api/merchant/orders/{id}:
    get:
      summary: 获取商家订单详情
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: 返回商家订单详情
...